Output rspec seed if used (#15)
Rspec supports running the specs in a [random order][1]. This helps in
surfacing any accidental dependencies and side effects in the test
suite. When debugging test failures caused by such side effects, it's
often necessary to reproduce the issue by running the specs in the same
order again. For that the seed value needs to be known and logged by the
formatter.

[1]: https://relishapp.com/rspec/rspec-core/v/3-8/docs/configuration/set-the-order-and-or-seed